Showcase credentials side by side
